% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{FigCorrelPlot}
\alias{FigCorrelPlot}
\title{Mapa de Correlação}
\usage{
FigCorrelPlot(Y, tipo = "Total", nfile = "output.png")
}
\arguments{
\item{Y}{Matriz de correlação.}

\item{tipo}{Tipo de correlação a ser mostrada: "Total" (parte inferior) ou "Parcial" (parte superior).}

\item{nfile}{Nome do arquivo de saída (PNG).}
}
\value{
Nenhum objeto retornado. Salva a figura em arquivo PNG.
}
\description{
Gera figura de correlação (com \code{corrplot}) para todas as variáveis.
}
\examples{
res <- correl_ind(mtcars)
FigCorrelPlot(res$Correl, tipo = "Total", nfile = "Correl.png")

}
