% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surpresa.R
\name{mostrar_foto}
\alias{mostrar_foto}
\title{Mostra a imagem de boas-vindas}
\usage{
mostrar_foto()
}
\value{
Uma imagem exibida no console gráfico.
}
\description{
Exibe uma imagem sempre que a função for chamada.
}
