% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacao.R
\name{multiplicacao}
\alias{multiplicacao}
\title{Multiplica dois números}
\usage{
multiplicacao(a, b)
}
\arguments{
\item{a}{Um número.}

\item{b}{Outro número.}
}
\value{
Produto de a e b
}
\description{
Esta função retorna a soma de dois números.
}
\examples{
soma(2, 3)
}
