% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Soma dois números}
\usage{
soma(a, b)
}
\arguments{
\item{a}{Um número.}

\item{b}{Outro número.}
}
\value{
A soma de a e b.
}
\description{
Esta função retorna a soma de dois números.
}
\examples{
soma(2, 3)
}
