% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-spp_rich.R
\name{spp_rich}
\alias{spp_rich}
\title{Calcula riqueza}
\usage{
spp_rich(x, ...)
}
\arguments{
\item{x}{A numeric vector of species abundances/presences in the community}

\item{...}{arguments to be passed to internal functions}
}
\value{
numeric
}
\description{
Essa função calcula a riqueza em um vetor de abundância/presença de
espécies de uma comunidade.
}
\examples{
mata <- c(3, 7, 10, 2, 0, 1)
spp_rich(mata)

}
